pylibs/chat/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "socratic-chat"
version = "0.1.0"
description = "Imperative chatting framework"
authors = ["Liu, Minsheng <lambda@liu.ms>"]
readme = "README.md"
packages = [
  { include = "socratic" }
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.340"
promptlayer = "^0.4.0"
python-dotenv = "^1.0.0"
openai = "^1.3.5"
pydantic = "^2.5.2"
aiochannel = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.342"
black = "^24.1.1"
isort = "^5.13.2"
pylint = "^3.0.3"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
async-timeout = "^4.0.3"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pyright]

[tool.pylint."messages control"]
disable = ["too-few-public-methods"]
ignore = ["setup.py", "__init__.py"]
ignore-paths = ["tests"]
min-similarity-lines=10

[tool.isort]
profile = "black"
line_length = 100
auto_identify_namespace_packages = false
force_single_line = true # pyright doesn't like implicit namespace + single line (related to https://github.com/microsoft/pyright/issues/2882?)
known_first_party = ["socratic"]
